[Fix] time integers are unsigned 32 bit

+++ src/fylgje-saver/DataManager.cpp
@@ -4,8 +4,6 @@
 #include "DataManager.h"
 
 
-
-
 int bifrost::data::hist_a_or_b(int x, int shift, int bins){
   int y = x >> shift;
   if (y < 0 || y >= bins) y = -1;
@@ -336,8 +334,6 @@ void bifrost::data::Manager::save_to(hdf5::node::Group group) const {
   message_dataset.write(messages);
 }
 
-
-
 hdf5::datatype::Compound bifrost::message_type() {
   auto compound = hdf5::datatype::Compound::create(sizeof(bifrost::message_t));
   compound.insert("fiber", offsetof(bifrost::message_t, fiber), hdf5::datatype::create<int>());
@@ -345,8 +341,8 @@ hdf5::datatype::Compound bifrost::message_type() {
   compound.insert("a", offsetof(bifrost::message_t, a), hdf5::datatype::create<int>());
   compound.insert("b", offsetof(bifrost::message_t, b), hdf5::datatype::create<int>());
   compound.insert("time", offsetof(bifrost::message_t, time), hdf5::datatype::create<double>());
-  compound.insert("high", offsetof(bifrost::message_t, high), hdf5::datatype::create<size_t>());
-  compound.insert("low", offsetof(bifrost::message_t, low), hdf5::datatype::create<size_t>());
+  compound.insert("high", offsetof(bifrost::message_t, high), hdf5::datatype::create<uint32_t>());
+  compound.insert("low", offsetof(bifrost::message_t, low), hdf5::datatype::create<uint32_t>());
   return compound;
 }
